separate header for type specific spline functions
