+++ cpp/splinepy/py/CMakeLists.txt
@@ -5,6 +5,7 @@ set(PYSPLINEPY_SRCS
         init/coordinate_reference.cpp
         init/core_spline.cpp
         init/fitting.cpp
+        init/spline_extensions.cpp
         splinepy_core.cpp)
 
 # add target
