+++ cpp/splinepy/py/init/core_spline.cpp
@@ -2,6 +2,6 @@
 
 namespace splinepy::py::init {
 namespace py = pybind11;
-void init_core_spline(py::module_& m) { add_spline_pyclass(m, "CoreSpline"); }
+void init_core_spline(py::module_& m) { add_spline_pyclass(m); }
 
 } // namespace splinepy::py::init
