+++ cpp/splinepy/py/splinepy_core.cpp
@@ -11,6 +11,9 @@ void init_coordinate_references(py::module_&);
 // CORE
 void init_core_spline(py::module_&);
 
+// Extensions
+void init_spline_extensions(py::module_& m);
+
 // Reader
 void init_reader(py::module_&);
 
@@ -29,6 +32,7 @@ namespace py = pybind11;
 PYBIND11_MODULE(splinepy_core, m) {
   splinepy::py::init::init_coordinate_references(m);
   splinepy::py::init::init_core_spline(m);
+  splinepy::py::init::init_spline_extensions(m);
   splinepy::py::init::init_reader(m);
   splinepy::py::init::init_exporter(m);
   splinepy::py::init::init_fitting(m);
